MAJOR ISSUE
I modified class Team and Match a lot. Every test and new tests passes.
However, I found out it will be hard to code function "becomes" in
Python, for the language doesn't allow to reassign an object passed as
the argument of a function... for example:

def becomes(winner, team):
    winner = team

does nothing to winner... what to do... see comments in code.