Refactored function Pool.ranking. Also added the ability to do a DoubleElimination tournament.
Refactored ranking using a new function in Team.
Added the LowerBracket class, that is used in the new DoubleElimination
class. These are used to create a Double elimination tournament.

The implementation of the LowerBracket is not perfect. There is a way to
make the match-ups better in that bracket (see the difference between
Challonge and us).

Added many TODOs in order to clearly state what is missing where. Could
be better but meh.

No tests were added.